issue #7 예외 처리2 - 정수가 아닌 실수를 입력할 시 - 입력된 실수도 계산이 가능하도록 x,y값을 실수로  처리

+++ calc/sum.py
@@ -9,7 +9,7 @@ def application(environ, start_response):
 	mul = 0
 	answer = ''
 	try:
-		x, y = int(x), int(y)
+		x, y = float(x), float(y)
 		sum = x + y
 		mul = x * y
 	except ValueError:
